Move View2/View3 to gui
